requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
colorama>=0.4.6
dnspython>=2.4.0
python-whois>=0.8.0
